requirements.txt
openai==1.3.7
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
